+++ integration-tests/tests/src/hooks/open-realm-before.ts
@@ -46,6 +46,9 @@ export function openRealmHook(config: LocalConfiguration | SyncedConfiguration =
       } as Realm.Configuration;
       this.realm = new Realm(this.config);
       // Upload the schema, ensuring a valid connection
+      if (!this.realm.syncSession) {
+        throw new Error("No syncSession found on realm");
+      }
       await this.realm.syncSession.uploadAllLocalChanges();
     }
   };
@@ -54,12 +57,16 @@ export function openRealmHook(config: LocalConfiguration | SyncedConfiguration =
 export function closeRealm(this: RealmContext): void {
   if (this.realm) {
     this.realm.close();
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore Ignore error as we ensure the realm is reopened
     delete this.realm;
   } else {
     throw new Error("Expected a 'realm' in the context");
   }
   if (this.config) {
     Realm.deleteFile(this.config);
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore Ignore error as we ensure the config is recreated
     delete this.config;
   } else {
     throw new Error("Expected a 'config' in the context");